requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.36
pymysql==1.1.1
alembic==1.13.2
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
celery==5.3.6
redis==5.0.8
slowapi==0.1.9
python-json-logger==2.0.7
watchdog==4.0.1
apscheduler==3.10.4
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
orjson==3.10.7
structlog==24.4.0
Jinja2==3.1.4
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0
factory-boy==3.3.0
freezegun==1.5.1
Faker==28.0.0
pandas==2.2.0
openpyxl==3.1.2
asyncpraw==7.8.1
hypothesis==6.98.3
